requirements.txt
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
watchdog>=3.0.0
typing-extensions>=4.0.0